refactor: add interface-folder
- add types to interface -folder
- change name of Item to ItemCard
