+++ client/src/Penalties.tsx
@@ -1,16 +1,11 @@
 import { Grid2 as Grid } from "@mui/material";
 import axios from "axios";
 import { useState, useEffect } from "react";
-import { Item } from "./components/Item";
+import { ItemCard } from "./components/ItemCard";
 import { useAuth } from "./auth/useAuth";
 import { Severity, useNotification } from "./store/NotificationContext";
 import { Loading } from "./layout/Loading";
-
-type Item = {
-  name: string,
-  price: number,
-  description?: string
-}
+import { Item } from '@putiikki/item'
 
 const API_URL = import.meta.env.VITE_API_URL;
 
@@ -61,7 +56,7 @@ export const Penalties = () => {
       {penalties.length > 0 && penalties.map((penalty: Item, index: number) => {
         return (
           <Grid size={{ lg: 4, md: 4, xs: 4, sm: 8 }} key={`grid-${index}`}>
-            <Item
+            <ItemCard
               item={penalty}
               key={`reward-${index}`}
               handlePoints={(points: number) => getPenalty(points)}
