+++ interfaces/user.d.ts
@@ -1,0 +1,10 @@
+export type User = {
+  name: string,
+  username: string,
+  points: number,
+}
+
+export type AuthData = {
+  username: string,
+  password: string,
+}
