+++ server/src/services/penalties.ts
@@ -1,22 +1,17 @@
+import { Item } from '@putiikki/item';
 import { getMany, update } from './db';
-
-type Penalty = {
-    name: string,
-    price: number,
-    description?: string
-}
 
 export async function getMultiple() {
     const data = await getMany(`SELECT * FROM penalties`);
 
-    const tasks: Penalty[] = data.map((item: Penalty) => ({
+    const tasks: Item[] = data.map((item: Item) => ({
         name: item.name, price: item.price, description: item.description
     }))
 
     return tasks;
 }
 
-export async function addPenalty(penalty: Penalty) {
+export async function addPenalty(penalty: Item) {
     const sql = `INSERT INTO penalties (name, price, description) VALUES ($1, $2, $3)`;
 
     await update(sql, [penalty.name, penalty.price, penalty.description]);
