+++ server/src/services/rewards.ts
@@ -1,22 +1,17 @@
 import { getMany, update } from './db';
-
-type Reward = {
-    name: string,
-    price: number,
-    description?: string
-}
+import { Item } from "@putiikki/item"
 
 export async function getMultiple(page = 1) {
     const data = await getMany(`SELECT * FROM rewards`);
 
-    const rewards: Reward[] = data.map((item: Reward) => ({
+    const rewards: Item[] = data.map((item: Item) => ({
         name: item.name, price: item.price, description: item.description
     }))
 
     return rewards;
 }
 
-export function addReward(reward: Reward) {
+export function addReward(reward: Item) {
     const sql = `INSERT INTO rewards (name, price, description) VALUES ($1, $2, $3)`;
     update(sql, [reward.name, reward.price, reward.description]);
 }
