+++ server/tsconfig.json
@@ -1,13 +1,23 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     "module": "es2020",
     "esModuleInterop": true,
     "target": "es6",
     "moduleResolution": "node",
     "sourceMap": true,
-    "outDir": "dist"
+    "outDir": "dist",
+    "paths": {
+      "@putiikki/*": [
+        "./../interfaces/*",
+        "./../../interfaces/*"
+      ],
+    }
   },
   "lib": [
     "es2015"
+  ],
+  "include": [
+    "src"
   ]
 }
